+++ crates/miden-objects/src/block/account_tree.rs
@@ -128,6 +128,13 @@ impl AccountTree {
         self.smt.root()
     }
 
+    /// Returns true if the tree contains a leaf for the given account ID prefix.
+    pub fn contains_account_id_prefix(&self, account_id_prefix: AccountIdPrefix) -> bool {
+        let key = Self::id_prefix_to_smt_key(account_id_prefix);
+        let is_empty = matches!(self.smt.get_leaf(&key), SmtLeaf::Empty(_));
+        !is_empty
+    }
+
     /// Returns the number of account IDs in this tree.
     pub fn num_accounts(&self) -> usize {
         // Because each ID's prefix is unique in the tree and occupies a single leaf, the number of
@@ -263,6 +270,16 @@ impl AccountTree {
         let mut key = Word::empty();
         key[Self::KEY_SUFFIX_IDX] = account_id.suffix();
         key[Self::KEY_PREFIX_IDX] = account_id.prefix().as_felt();
+
+        key
+    }
+
+    /// Returns the SMT key of the given account ID prefix.
+    fn id_prefix_to_smt_key(account_id: AccountIdPrefix) -> Word {
+        // We construct this in such a way that we're forced to use the constants, so that when
+        // they're updated, the other usages of the constants are also updated.
+        let mut key = Word::empty();
+        key[Self::KEY_PREFIX_IDX] = account_id.as_felt();
 
         key
     }
@@ -484,5 +501,23 @@ pub(super) mod tests {
             assert_eq!(witness.leaf(), control_leaf);
             assert_eq!(witness.path(), &control_path);
         }
+    }
+
+    #[test]
+    fn contains_account_prefix() {
+        // Create a tree with a single account.
+        let [pair0, pair1] = setup_duplicate_prefix_ids();
+        let tree = AccountTree::with_entries([(pair0.0, pair0.1)]).unwrap();
+        assert_eq!(tree.num_accounts(), 1);
+
+        // Validate the leaf for the inserted account exists.
+        assert!(tree.contains_account_id_prefix(pair0.0.prefix()));
+
+        // Validate the leaf for the uninserted account with same prefix exists.
+        assert!(tree.contains_account_id_prefix(pair1.0.prefix()));
+
+        // Validate the unrelated, uninserted account leaf does not exist.
+        let id1 = AccountIdBuilder::new().build_with_seed([7; 32]);
+        assert!(!tree.contains_account_id_prefix(id1.prefix()));
     }
 }